Make is_virtual part of contribution to common category
Since is_virtual is not a default category it requires that it is defined at the site level in binder_config.yaml. To work around that requirement, the ntp module can make the decisions and contribute the result in the common category.

This commit adds two paths to the common category contribution: for data found in 'data/is_virtual/true.{yaml, json}' and the corresponding files called 'false'.